Move `state` utility into `store/util`
