+++ src/sidebar/store/util.js
@@ -1,3 +1,5 @@
+/** @typedef {import('redux').Store} Store */
+
 /**
  * Return an object where each key in `updateFns` is mapped to the key itself.
  *
@@ -65,4 +67,32 @@ export function bindSelectors(namespaces, getState) {
     });
   });
   return boundSelectors;
+}
+
+/**
+ * Return a value from app state when it meets certain criteria.
+ *
+ * `await` returns a Promise which resolves when a selector function,
+ * which reads values from a Redux store, returns non-null.
+ *
+ * @template T
+ * @param {Object} store - Redux store
+ * @param {(s: Store) => T|null} selector - Function which returns a value from the
+ *   store if the criteria is met or `null` otherwise.
+ * @return {Promise<T>}
+ */
+export function awaitStateChange(store, selector) {
+  const result = selector(store);
+  if (result !== null) {
+    return Promise.resolve(result);
+  }
+  return new Promise(resolve => {
+    const unsubscribe = store.subscribe(() => {
+      const result = selector(store);
+      if (result !== null) {
+        unsubscribe();
+        resolve(result);
+      }
+    });
+  });
 }
